Add more help text and verbose option.
Add the verbose option to turn off most output by default.
Otherwise, the amount of output is controlled by the OMP_NUM_THREADS
setting, which makes the runs look different for 1 thread vs. more than one thread.

The verbose output is still disabled when there is more than one thread.
In the future the output could be restricted to to a single thread, and lift this restriction.