Fix compares with unsigned variables.
The comparison for a negative value with an unsigned type will always be false,
so we can remove the code.  I think it is safe to assume that all systems have
an unsigned size_t now.
In the RNG seeding case, check for 0 instead of a negative value (the analogous
code in BoostRandom checks for <= 0)