Fix Jastrow initialization.
The Jastrow values in the Input files are parameter values.
The parameter values are almost the spline coefficients - they need to be
combined with the cusp condition to get the final spline coefficients.
The values passed to the 'initialize' method are the actual Jastrow function values
that are fit to get parameter values.

Create a new method ('setupParameters') to initialize the Jastrow functor.
Remove the 'initialize' method and the LinearFit code that it uses.