Write timing information to an xml file.
It uses the same format as the QMCPACK info.xml file.
The XML handling is done using the TinyXML2 library, which is included in this commit and adds no extra external dependencies.

The file is named with the tiling (e.g. info_1_1_1.xml) to make it easier for the results of differently sized runs to coexist.

The output is enhanced with respect to the QMCPACK info file:
 - The particle size information is output (number of electrons).  The same output should be in the QMCPACK info file soon.
 - The driver name and number of steps and substeps. Similiar information should go into the QMCPACK info file, but the form may differ.

The hardware section is not yet populated (number of OMP threads, etc.)