Intel Compiler: workaround if libomp-dev is installed on system.
For intel offload targets, OpenMP initialization is stuck in hanging state
if the user has 'libomp-dev' package installed on system. The package
adds a symbolic link 'libiomp5.so -> libomp.so'.

The MKL detection from CMake picks up this symbolic link and adds the
library to the link line for the miniqmc application.

This patch implements a workaround to just incorporate oneAPI's Library
environment into CMake's to supersede the system libraries.